requirements.txt
Pillow==10.0.0
Pillow==10.0.0
Requests==2.31.0
selenium==4.10.0
